+++ tests/test_skip_remote_data.py
@@ -2,8 +2,6 @@
 # this test doesn't actually use any online data, it should just be skipped
 # by run_tests because it has the remote_data decorator.
 
-import sys
-
 import pytest
 from contextlib import closing
 
@@ -12,6 +10,8 @@ from six.moves.urllib.request import urlopen
 
 
 ASTROPY_DATA_URL = "http://data.astropy.org/"
+GITHUB_DATA_URL = "http://astropy.github.io/"
+EXTERNAL_URL = "http://www.google.com"
 TIMEOUT = 10
 
 if six.PY2:
@@ -25,6 +25,7 @@ def download_file(remote_url):
     with closing(urlopen(remote_url, timeout=TIMEOUT)) as remote:
         remote.read()
 
+
 @pytest.mark.remote_data
 def test_skip_remote_data(pytestconfig):
 
@@ -35,12 +36,17 @@ def test_skip_remote_data(pytestconfig):
         pytest.fail('@remote_data was not skipped with remote_data=none')
     elif pytestconfig.getoption('remote_data') == 'astropy':
         pytest.fail('@remote_data was not skipped with remote_data=astropy')
+    elif pytestconfig.getoption('remote_data') == 'github':
+        pytest.fail('@remote_data was not skipped with remote_data=github')
 
     # Test Astropy URL
     download_file(ASTROPY_DATA_URL + 'galactic_center/gc_2mass_k.fits')
 
-    # Test non-Astropy URL
-    download_file('http://www.google.com')
+    # Test GitHub URL
+    download_file(GITHUB_DATA_URL)
+
+    # Test unrelated URL
+    download_file(EXTERNAL_URL)
 
 
 @pytest.mark.remote_data(source='astropy')
@@ -49,8 +55,9 @@ def test_skip_remote_data_astropy(pytestconfig):
     # astropy.test() has remote_data=none but we still got here somehow,
     # so fail with a helpful message
 
-    if pytestconfig.getoption('remote_data') == 'none':
-        pytest.fail('@remote_data was not skipped with remote_data=none')
+    if pytestconfig.getoption('remote_data') in ('none', 'github'):
+        pytest.fail('@remote_data was not skipped with remote_data=none'
+                    'or remote_data=github')
 
     # Test Astropy URL
     download_file(ASTROPY_DATA_URL + 'galactic_center/gc_2mass_k.fits')
@@ -58,10 +65,27 @@ def test_skip_remote_data_astropy(pytestconfig):
     # Test non-Astropy URL
     if pytestconfig.getoption('remote_data') == 'astropy':
         with pytest.raises(Exception) as exc:
-            download_file('http://www.google.com')
+            download_file(EXTERNAL_URL)
+        assert "An attempt was made to connect to the internet" in str(exc.value)  # noqa
+    else:  # remote_data=any
+        download_file(EXTERNAL_URL)
+
+
+@pytest.mark.remote_data(source='github')
+def test_skip_remote_data_github(pytestconfig):
+    if pytestconfig.getoption('remote_data') == 'none':
+        pytest.fail('@remote_data was not skipped with remote_data=none')
+
+    # Test GitHub URL
+    download_file(GITHUB_DATA_URL)
+
+    # Test non-GitHub URL
+    if pytestconfig.getoption('remote_data') == 'github':
+        with pytest.raises(Exception) as exc:
+            download_file(ASTROPY_DATA_URL)
         assert "An attempt was made to connect to the internet" in str(exc.value)
-    else:
-        download_file('http://www.google.com')
+    else:  # remote_data=any or remote_data=astropy
+        download_file(ASTROPY_DATA_URL)
 
 
 @pytest.mark.internet_off
@@ -74,15 +98,18 @@ def test_internet_off_decorator(pytestconfig):
 def test_block_internet_connection(pytestconfig):
     if pytestconfig.getoption('remote_data') == 'none':
         with pytest.raises(_EXPECTED_ERROR):
-            download_file('http://www.google.com')
+            download_file(EXTERNAL_URL)
     elif pytestconfig.getoption('remote_data') == 'astropy':
         with pytest.raises(_EXPECTED_ERROR):
-            download_file('http://www.google.com')
+            download_file(EXTERNAL_URL)
+    elif pytestconfig.getoption('remote_data') == 'github':
+        with pytest.raises(_EXPECTED_ERROR):
+            download_file(EXTERNAL_URL)
     else:
-        download_file('http://www.google.com')
+        download_file(EXTERNAL_URL)
 
 
 @pytest.mark.internet_off
 def test_block_internet_connection_internet_off():
     with pytest.raises(_EXPECTED_ERROR):
-        download_file('http://www.google.com')
+        download_file(EXTERNAL_URL)